+++ data-clist.cabal
@@ -12,7 +12,7 @@ Author: John Van Enk <vanenkj@gmail.com>
 Maintainer: Jeremy Huffman <jeremy@jeremyhuffman.com>, John Van Enk <vanenkj@gmail.com>
 Stability: experimental
 Category: Data Structures
-Cabal-Version: >= 1.6
+Cabal-Version: >= 1.8
 Build-Type: Simple
 Homepage: https://github.com/sw17ch/data-clist
 
@@ -22,11 +22,19 @@ source-repository head
 
 Library
     Build-Depends: base >= 4 && < 5,
-                   deepseq >= 1.1 && < 1.5,
-                   QuickCheck >= 2.4 && < 2.15
-
+                   deepseq >= 1.1 && < 1.5
     Exposed-Modules:
         Data.CircularList
+        Data.CircularList.Internal
 
     ghc-options:        -Wall
     hs-source-dirs:     src
+
+Test-Suite tests
+  Type:             exitcode-stdio-1.0
+  Build-Depends:    base >=4.11 && < 5
+                  , data-clist
+                  , QuickCheck >= 2.4 && < 2.15
+  hs-source-dirs:   tests/
+  main-is:          quickcheck.hs
+
